dont require mocking script inside tests
